refactor: Lotto에 종속되는 상수들의 위치 변경 및 예외 추가
