+++ src/main/java/lotto/model/lotto/generator/RandomLottoGenerator.java
@@ -1,8 +1,11 @@
 package lotto.model.lotto.generator;
 
+import static camp.nextstep.edu.missionutils.Randoms.pickUniqueNumbersInRange;
 import static java.util.Comparator.naturalOrder;
+import static lotto.model.lotto.Lotto.END_INCLUSIVE;
+import static lotto.model.lotto.Lotto.MAX_NUMBER_COUNT;
+import static lotto.model.lotto.Lotto.START_INCLUSIVE;
 
-import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.LongStream;
@@ -10,10 +13,6 @@ import lotto.model.lotto.Lotto;
 import lotto.model.lotto.Lottos;
 
 public class RandomLottoGenerator implements LottoGenerator {
-
-    private static final int START_INCLUSIVE = 1;
-    private static final int END_INCLUSIVE = 45;
-    private static final int COUNT = 6;
 
     @Override
     public Lottos generate(final int lottoCount) {
@@ -27,7 +26,7 @@ public class RandomLottoGenerator implements LottoGenerator {
     }
 
     private List<Integer> generateSortedUniqueNumbers() {
-        return Randoms.pickUniqueNumbersInRange(START_INCLUSIVE, END_INCLUSIVE, COUNT)
+        return pickUniqueNumbersInRange(START_INCLUSIVE, END_INCLUSIVE, MAX_NUMBER_COUNT)
                 .stream()
                 .sorted(naturalOrder())
                 .toList();
