+++ src/test/java/lotto/io/validator/lotto/LottoLengthValidatorTest.java
@@ -3,7 +3,7 @@ package lotto.io.validator.lotto;
 import static lotto.io.error.ErrorMessage.INVALID_LOTTO_LENGTH_RANGE;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import lotto.io.exception.InvalidRangeException;
+import lotto.exception.InvalidRangeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
